refactor(dev): expose dev app and entrypoint
