+++ packages/astro/src/vite-plugin-astro-server/index.ts
@@ -1,3 +1,2 @@
-export { DevApp } from './app.js';
 export { createController, runWithErrorHandling } from './controller.js';
 export { default as vitePluginAstroServer } from './plugin.js';
